Added Document type
I noticed that simd-json was still incredibly fast even without simd
enabled. The Document type is my last effort at making a mostly-safe
JSON parser.

The Document type parses into a single Vec, which is advantageous for
nested payloads due to having less tiny allocations. The downside is
that it takes a little more work to traverse the parsed document than
the Value type.

I've implemented a lot of convenience methods to facilitate moving
between Documents into Values, since the Value type is the only type
that supports being transformed back to JSON currently.

The major change in this commit is moving the parsing logic into its own
module and sharing it between the Document and Value types. This also
led to me publishing that event-driven parser module.