Fixed bug with invalid utf-8 escape sequence
When attempting to create a std::str::Utf8Error, I made a mistake in
trying to convert the decoded representation of the bytes into a string.
It turns out that the codepoint \udddd in its decoded form *happens* to
be a valid uniode sequence when interpreted as bytes. By changing the
unit test to a different invalid codepoint, we can trigger a panic in
the original code.

The fix was already something on my todo list: remove the Utf8Error.
Since the Error type already has the offset of the invalid unicode
codepoint, having the original error wasn't needed. And, it allows this
problematic code to be removed.

Thanks to @pinkforest for reporting this and providing a simple fuzzer
for me to wake up to with my morning coffee!