JsonString now can be a raw string
The last set of changes made it possible to update values in-place, but
JsonString::from(&str) would require allocating a new string if it
contained any escapes. Now, the JSON-encoding process is done lazily,
just like the decoding process is done lazily.

This should amount to fewer allocations when editing an existing Value
and re-rendering to JSON.