Simplified surrogate detection
Originally I thought I was going to need more complex logic, and the
match statement was added. After reviewing the full commit, there was no
point in me inlining the logic from char::from_u32 -- I had originally
thought I could combine multiple ranges in a single match, but then
never ended up doing that.

Now, we have the optimized code flow for from_u32, and if a surrogate is
encountered, it's handled afterwards. This could technically be a little
slower than before for the surrogate codepoints themselves, but I can't
imagine the extra u32 checks will be that impactful.