Raw strings now promote to Json when possible
This set of changes adds more convenience functions to JsonString. It is
now possible to get access to the internal string buffer either as
encoded or decoded, and receive None if the operation isn't possible.

The other major change is that on conversion from String/str for
JsonString, the string is checked for characters that will need to be
escaped. The expected encoded length is measured, and if any escapes are
needed, the string is kept in raw form with the extra information to
allow exact-size allocations when encoding.

This also enables more quick-checks that can avoid full string
comparisons when comparing to JsonStrings -- if one JsonString is raw
and the other is encoded, the expected length can be checked before
committing to the more expensive comparison.