Expanding coverage of String and Document parsing
The json.org test suite now calls both the bytes and str parsing methods
on all types. This is slightly overkill, but it's a flow that explicitly
tests a custom config, and that was where one aspect of coverage was
missing -- Document::from_json_with_config. By making this suite test
all the parsing methods, it ensures we have good coverage in a way that
tests that the config is actually being used.

Additionally, JsonString::is_empty was added but had nothing testing it.