Added Value::as_str plus others
Closes #9

The as_str changes give easier access to a &str reference, but since
JsonStrings are lazily-decoded, this means None will be returned if any
escape sequences are contained. In general, extracting an &str is a bit
of an anti-pattern because of this and because JsonString is designed to
be able to be compared against strings without that limitation.

Additionally, JsonString now implements Display, making the standard
ToString convert the JsonString to its decoded format.